+++ src/interfaces/IPosition.h
@@ -3,7 +3,7 @@
  * @author Sniehovskyi Nikita (xsnieh00)
  * @author TODO
  * @date 08.04.2024
- * @brief Definition of Position interface
+ * @brief Definition of Position doubleerface
  */
 
 #ifndef __IPOSITION_H__
@@ -11,13 +11,13 @@
 
 class IPosition {
     public:
-        int x;
-        int y;
-        int angle;
+        double x;
+        double y;
+        double angle;
 
         virtual void SetPosition(IPosition * position) = 0;
-        virtual void SetPosition(int x, int y, int angle) = 0;
-        virtual void SetPosition(int x, int y) = 0;
+        virtual void SetPosition(double x, double y, double angle) = 0;
+        virtual void SetPosition(double x, double y) = 0;
         virtual ~IPosition() {};
 };
