+++ src/interfaces/IRobot.h
@@ -3,7 +3,7 @@
  * @author Sniehovskyi Nikita (xsnieh00)
  * @author TODO
  * @date 08.04.2024
- * @brief Definition of Robot interface
+ * @brief Definition of Robot doubleerface
  */
 
 #ifndef __IROBOT_H__
@@ -14,15 +14,15 @@
 class IRobot : public IGameObject {
     public:
         virtual void Update() = 0;
-        virtual int GetSpeed() = 0;
-        virtual void SetSpeed(int speed) = 0;
-        virtual int GetRotationAngle() = 0;
-        virtual void SetRotationAngle(int rotationAngle) = 0;
+        virtual double GetSpeed() = 0;
+        virtual void SetSpeed(double speed) = 0;
+        virtual double GetRotationAngle() = 0;
+        virtual void SetRotationAngle(double rotationAngle) = 0;
         virtual ~IRobot() {};
 
     protected:
-        int _speed; // Constant through run
-        int _rotationAngle; // Constant through run
+        double _speed; // Constant through run
+        double _rotationAngle; // Constant through run
 };
 
 #endif
